+++ math/src/main/scala/au/id/cxd/text/count/TermCountTypeAliases.scala
@@ -2,6 +2,6 @@ package au.id.cxd.text.count
 
 object TermCountTypeAliases {
   type TermHashCode = Int
-  type TermCount = Double
+  type TermCount = Int
   type TermDocumentCount = Int
 }
